+++ src/common/utils/index.js
@@ -1,5 +1,6 @@
 export * from './fetchLocations';
 export * from './getDehydratedPiece';
 export * from './getHydratedPieces';
-export * from './ifMissingPermanentLoanTypeId';
+export * from './handleCommonErrors';
+export * from './handleReceiveErrorResponse';
 export * from './receive';